requirements.txt
networkx==1.7
sparql-client==0.13
wsgiref==0.1.2
